Move scorers and strategy to logic/
To compartmentalize
