+++ game/round.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/joshprzybyszewski/cribbage/model"
-	"github.com/joshprzybyszewski/cribbage/pegging"
+	"github.com/joshprzybyszewski/cribbage/logic/pegging"
 )
 
 const (
