+++ logic/scorer/describer.go
