+++ logic/scorer/scorer_test.go
