+++ logic/strategy/qa/main.go
@@ -8,7 +8,7 @@ import (
 	survey "github.com/AlecAivazis/survey/v2"
 
 	"github.com/joshprzybyszewski/cribbage/model"
-	"github.com/joshprzybyszewski/cribbage/strategy"
+	"github.com/joshprzybyszewski/cribbage/logic/strategy"
 )
 
 func main() {
