+++ logic/strategy/simple_crib.go
