+++ logic/strategy/simple_peg.go
