+++ logic/strategy/simple_peg_test.go
